Python/pyproject.toml
[project]
name = "LinkedIn"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.128.7",
    "google-cloud-firestore>=2.23.0",
    "ipykernel>=7.2.0",
    "google-genai>=1.0.0",
    "pandas>=3.0.0",
    "python-dotenv>=1.0.0",
    "uvicorn>=0.40.0",
]
